Refactored and reorganized feature.
