+++ client/src/pages/Login/Login.tsx
@@ -6,58 +6,11 @@ import { useSnackBar } from '../../context/useSnackbarContext';
 import PageContainer from '../../components/PageContainer/PageContainer';
 import AuthPageWrapper from '../../components/AuthPageWrapper/AuthPageWrapper';
 import AuthPageFooter from '../../components/AuthPageFooter/AuthPageFooter';
-import Button from '@mui/material/Button';
-import { FetchOptions } from '../../interface/FetchOptions';
-import useStyles from './useStyles';
-
-function DemoLogin(): JSX.Element {
-  const { updateLoginContext } = useAuth();
-  const { updateSnackBarMessage } = useSnackBar();
-  const classes = useStyles();
-
-  const handleSubmit = async () => {
-    const fetchOptions: FetchOptions = {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    };
-    return await fetch(`/auth/demo/login`, fetchOptions)
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.error) {
-          updateSnackBarMessage(data.error.message);
-        } else if (data.success) {
-          updateSnackBarMessage(data.success);
-          updateLoginContext(data.success);
-        } else {
-          console.log({ data });
-          updateSnackBarMessage('unexpected error occurred!');
-        }
-      })
-      .catch(() => ({
-        error: { message: 'Unable to connect to server. Please try again' },
-      }));
-  };
-  return (
-    <form onSubmit={handleSubmit}>
-      <Button
-        className={classes.demoButton}
-        type="submit"
-        size="large"
-        variant="contained"
-        color="primary"
-        disableElevation
-      >
-        Demo Login
-      </Button>
-    </form>
-  );
-}
+import DemoButton from '../../components/DemoButton/DemoButton';
 
 export default function Login(): JSX.Element {
   const { updateLoginContext } = useAuth();
   const { updateSnackBarMessage } = useSnackBar();
-
   const handleSubmit = (
     { email, password }: { email: string; password: string },
     { setSubmitting }: FormikHelpers<{ email: string; password: string }>,
@@ -82,7 +35,7 @@ export default function Login(): JSX.Element {
     <PageContainer>
       <AuthPageWrapper header="Log in">
         <LoginForm handleSubmit={handleSubmit} />
-        <DemoLogin />
+        <DemoButton text="Demo Login" path="login" />
         <AuthPageFooter text="Not a member?" anchorText="Sign up" anchorTo="/signup" />
       </AuthPageWrapper>
     </PageContainer>
