+++ client/src/pages/Login/useStyles.ts
@@ -18,19 +18,6 @@ const useStyles = makeStyles((theme: Theme) => ({
     color: '#000000',
     fontWeight: 700,
   },
-  demoButton: {
-    margin: theme.spacing(3, 2, 2),
-    padding: 10,
-    width: 160,
-    height: 56,
-    borderRadius: theme.shape.borderRadius,
-    fontSize: 16,
-    backgroundColor: theme.palette.primary.main,
-    fontWeight: 'bold',
-    textTransform: 'uppercase',
-    left: 95,
-    top: 15,
-  },
 }));
 
 export default useStyles;
